+++ kudzu-core/src/commonMain/kotlin/com/copperleaf/kudzu/parser/expression/ExpressionEvaluator.kt
@@ -5,10 +5,11 @@ import com.copperleaf.kudzu.node.Node
 /**
  * Evaluates the parse-tree created by [ExpressionParser].
  */
+@ExperimentalStdlibApi
 interface ExpressionEvaluator<T : Any> {
 
     /**
      * Evaluate the expression starting at [Node], which should be the result of [ExpressionParser.parse].
      */
-    fun evaluateExpression(node: Node): T
+    val evaluate: DeepRecursiveFunction<Node, T>
 }